backend/linepsite/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.1.7
django-cors-headers==3.7.0
django-getenv==1.3.2
djangorestframework==3.12.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
requests==2.25.1
python-socketio==5.3.0
eventlet==0.31.0
channels==3.0.3
uvicorn[standard]==0.14.0
smbprotocol==1.5.1
gunicorn==20.1.0
debugpy==1.5.1
pyodbc==4.0.34

